Force write in conf

+++ worker_node/conf.py
@@ -25,16 +25,19 @@ class Config:
         return "%s/%s.%s" % (cls.path, name, file_type)
 
     @classmethod
-    def save(cls, name, data, file_type="json"):
+    def save(cls, name, data, file_type="json", force=False):
         """
         Save configurations to a file
         :param name: name of the config
         :param data: data to be stored
         :param file_type: config type (default: json)
+        :param force: force write
         :return: path of the file
         """
         cls.__build_path()
         path = cls.get_path(name, file_type)
+        if os.path.isfile(path) and not force:
+            return path
         if file_type == "json":
             data = dumps(data, indent=4)
         fl = open(path, 'w')
